moved theme into the ui package
